refactor: Add `TemplateCompiler` interface, make implementation final
There should be no need to extend the logic of the stock template
compiler, the only real need is to be able to stub / mock it for tests.

Move the actual implementation to `CoreTemplateCompiler` and introduce
an interface for use elsewhere.
